requirements.txt
Django==1.11.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
Pillow==5.3.0
pytz==2018.7
six==1.11.0
